+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
-## [1.0.16-cloud] - 2023-12-24
+## [1.0.16-cloud] - 2023-12-31
 ### Added
 - Add 'short' param to list resources api